Put syncDataWithServer function in rofService
